Clipboard redesign: Implement the core functionality
This change introduces the new clipboard framework. This includes the
following features:

Selection and pasting have been decoupled. In the previous
implementation, panes that inherited from CUT-AND-PASTE-MIXIN would do
two things:

  1) If the pane was a stream pane, it was possible to shift-drag to
     select text. After selection, the implementation took ownership
     of the primary selection. The API's used to do this were highly
     dependent on the selection mechanism in X11.

  2) Any shift-click events would be stolen, and shift-middle-click
     would result in a CONVERT-SELECTION command being sent to the
     X-server (again, the API to do this was highly dependent on X11).

With the new implementation, only the selection part remains and the
mixin has been renamed to SELECTION-MIXIN in order to better represent
what it does. Handling of shift-middle-click is now left to the
specific panes that are interested in this behaviour. This will be
implemented in the next commit.

The new framework introduces support for both selections and the
clipboard. The selection is the content of text that has been
highlighted using shift-drag. An application can assign some value to
the selection using the function COPY-TO-SELECTION. In X11, most text
fields allows you to paste the content of the selection by
middle-clicking.

The clipboard is another way to transfer information between
applications. In most X11 applications, the content of the clipboard
is pasted using Control-V. The content of the clipboard can be set
using the function COPY-TO-CLIPBOARD.

Finally, this patch implements the ability to request information from
the selection and clipboard. This is done using the functions
REQUEST-SELECTION-CONTENT and REQUEST-CLIPBOARD-CONTENT. After these
functions are called, an event will be delivered which contains the
content, if it was available.